requirements.txt
ollama
streamlit
pdf2image
pytesseract
pillow